Change edit and delete links into icons

+++ src/main/webapp/WEB-INF/view/list-budgets.jsp
@@ -12,6 +12,16 @@
 <title>List Budget Items</title>
 <%@include file="head.jsp"%>
 
+<link rel="stylesheet"
+	href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
+	integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO"
+	crossorigin="anonymous">
+
+<link rel="stylesheet"
+	href="https://use.fontawesome.com/releases/v5.3.1/css/all.css"
+	integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU"
+	crossorigin="anonymous">
+
 </head>
 <body>
 
@@ -144,10 +154,10 @@
 						<th class="col-xs-1" scope="col" index=3>% of Usage
 							<div class="filter"></div>
 						</th>
-						<th class="col-xs-3" scope="col" index=4>Description
+						<th class="col-xs-4" scope="col" index=4>Description
 							<div class="filter"></div>
 						</th>
-						<th class="col-xs-3" scope="col">Action</th>
+						<th class="col-xs-2" scope="col">Action</th>
 					</tr>
 				</thead>
 				<tbody>
@@ -183,11 +193,14 @@
 							
 							<td id="percent" class="col-xs-1"></td>
 									
-							<td class="col-xs-3">${budgetItem.budgetDetailDescription}</td>
-							<td class="col-xs-3"><a class="updateLink" href="${updateLink}">Update</a> | 
-							<a class="updateLink" href="${deleteLink}"
-								onclick="if(!(confirm('Are you sure you want to delete this item?'))) return false">Delete</a>
-							</td>
+							<td class="col-xs-4">${budgetItem.budgetDetailDescription}</td>
+							<td class="col-xs-2">
+							
+							<a class="updateLink" href="${updateLink}" title="Edit"><i class="fas fa-edit"></i></a> | 
+							<a class="updateLink" href="${deleteLink}" title="Delete" onclick="if(!(confirm('Are you sure you want to delete this item?'))) return false">
+							<i class="fas fa-trash-alt"></i></a></td>
+							
+							
 						</tr>
 
 					</c:forEach>
